# docs/api/openapi.yaml
...
openapi: 3.0.3
info:
  title: Hamlaneh API
  description: >
    Self-hosted team communication platform API. Phase 1.1 surface: health,
    authentication, current user, and the admin user-management endpoints that
    back the dashboard (public registration is off by default; users are
    created from the admin dashboard or via invite links). Phase 1.2 adds the
    messaging surface the chat shell needs: channels and 1:1 direct messages
    with membership, cursor-paged message history with idempotent send, edit
    and soft delete, per-user read positions, message search, and the
    WebSocket upgrade that carries realtime delivery. Phase 1.1b completes the
    identity surface: email password reset, TOTP two-step verification with
    recovery codes, and self-service session management.
  version: 0.6.0
  license:
    name: AGPL-3.0
    url: https://www.gnu.org/licenses/agpl-3.0.html
servers:
  - url: /
    description: Same-origin deployment behind Caddy.
...
paths:
...
  /api/v1/auth/login/totp:
    post:
      tags: [auth]
      operationId: completeTotpLogin
      summary: Complete a two-step sign-in with an authenticator or recovery code.
      description: >
        The second half of login. Authenticated by the hamlaneh_2fa challenge
        cookie the 202 login set — deliberately no security scheme here,
        mirroring /auth/refresh; the authz matrix classifies it as
        challenge-cookie-gated. `code` is the six-digit authenticator code or
        one unused recovery code (case-insensitive, hyphens optional); a
        recovery code is consumed by the attempt that uses it. An accepted
        authenticator code is never accepted twice, even inside its
        thirty-second window. Five wrong codes revoke the challenge and the
        caller starts again at the password step. On success the challenge is
        consumed and session cookies are set exactly as a plain login sets them.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TotpLoginRequest"
      responses:
        "200":
          description: Authenticated; session cookies set.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          description: >
            No live challenge — missing, expired, consumed or revoked (code
            not_authenticated; return to the password step) — or a live
            challenge with a wrong code (code invalid_totp_code; the cells
            clear and the caller retries).
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "429":
          $ref: "#/components/responses/RateLimited"
...
  /api/v1/users/me/totp/setup:
    post:
      tags: [users]
      operationId: startTotpSetup
      summary: Start (or restart) two-step verification setup.
      description: >
        Generates a fresh 160-bit secret and returns the manual key, the
        otpauth URI and a server-rendered QR (the settings handoff: rendered
        server-side from the secret, never a stored image). The pending setup
        lives one hour; calling again before activation replaces it, which is
        also why Cancel needs no endpoint. The account stays password-only
        until activate. Answers carry Cache-Control: no-store.
      security:
        - sessionCookie: []
      responses:
        "200":
          description: Pending setup created or replaced.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TotpSetup"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "409":
          description: Two-step verification is already on (code totp_already_enabled).
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "429":
          $ref: "#/components/responses/RateLimited"
  /api/v1/users/me/totp/verify:
    post:
      tags: [users]
      operationId: verifyTotpSetup
      summary: Verify the first authenticator code (setup step 2).
      description: >
        Checks a six-digit code against the pending secret. A wrong code does
        not restart setup — the secret stays valid and the client clears the
        cells — but five wrong codes revoke the pending setup, because an
        uncapped verifier is a brute-force oracle. On success the ten recovery
        codes are generated and returned, shown exactly once (only argon2id
        hashes are kept). Two-step verification is still OFF until activate.
        Answers carry Cache-Control: no-store.
      security:
        - sessionCookie: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/VerifyTotpSetupRequest"
      responses:
        "200":
          description: Code accepted; the recovery codes for step 3.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RecoveryCodes"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          description: Wrong code (code invalid_totp_code); the setup survives.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "409":
          description: >
            No pending setup, it expired, or the attempt cap revoked it (code
            totp_setup_expired) — start again at step 1.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "429":
          $ref: "#/components/responses/RateLimited"
  /api/v1/users/me/totp/activate:
    post:
      tags: [users]
      operationId: activateTotp
      summary: Turn two-step verification on (setup step 3).
      description: >
        The action behind the saved-the-codes acknowledgement. Only a VERIFIED
        pending setup activates: the server never turns two-step verification
        on for an account that has not proven its authenticator and been shown
        its recovery codes, so closing the panel at step 3 can never strand an
        account whose only fallback was never displayed.
      security:
        - sessionCookie: []
      responses:
        "204":
          description: Two-step verification is on for the next sign-in.
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "409":
          description: >
            Nothing to activate — no pending setup, not yet verified, or
            expired (code totp_setup_not_verified).
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
...
  /api/v1/users/me/oidc:
    post:
      tags: [users]
      operationId: linkOidcIdentity
      summary: Connect single sign-on to this account.
      description: >
        Answers with the provider redirect the browser should follow, and sets
        the same transaction cookie the sign-in flow uses, carrying the intent
        to link and the id of the signed-in account. The callback then records
        the identity against that account instead of minting a session.

        Binding the account id into a server-minted HttpOnly cookie, and the
        state into both the cookie and the URL, is what makes link fixation
        fail: the transaction exists only in the browser that started it.
      security:
        - sessionCookie: []
      responses:
        "200":
          description: Follow this to the provider.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OidcRedirect"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "409":
          description: >
            This account already has an identity linked (code
            sso_already_linked).
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "429":
          $ref: "#/components/responses/RateLimited"
        "503":
          description: Single sign-on is not configured (code sso_unavailable).
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
    delete:
      tags: [users]
      operationId: unlinkOidcIdentity
      summary: Disconnect single sign-on from this account.
      description: >
        Refused when the account has no password, because unlinking would
        leave it with no way in at all. The recovery path for such an account
        is an administrator issuing a temporary password, which is also the
        break-glass door when the provider is unreachable.
      security:
        - sessionCookie: []
      responses:
        "204":
          description: The identity is no longer linked.
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          description: Nothing was linked (code sso_not_linked).
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "409":
          description: >
            The account has no password, so this is its only way in (code
            sso_unlink_no_password).
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "429":
          $ref: "#/components/responses/RateLimited"
  /api/v1/users/me/totp/disable:
    post:
      tags: [users]
      operationId: disableTotp
      summary: Turn two-step verification off.
      description: >
        Re-asks for the password (the design's confirm dialog). Removes the
        secret and invalidates every recovery code. Sessions are deliberately
        NOT revoked: the threat is a hijacked session entrenching itself, and
        revoking other families would punish only the legitimate user's
        devices while the attacker's own session survives. The password
        prompt, the rate limit and the audit log are the defences.

        Refused outright while the organisation requires two-step
        verification. Enforcement binds when a session is minted, so without
        this an account could enrol, sign in unflagged, switch the factor
        back off and keep refreshing — the policy would hold for everyone
        except the people it is aimed at.
      security:
        - sessionCookie: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PasswordConfirmRequest"
      responses:
        "204":
          description: Two-step verification is off; recovery codes void.
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          description: >
            Password incorrect (code invalid_current_password), or the
            organisation requires two-step verification (code
            totp_required_by_org).
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "409":
          description: Two-step verification is not on (code totp_not_enabled).
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "429":
          $ref: "#/components/responses/RateLimited"
...
  /api/v1/users:
    get:
      tags: [users]
      operationId: listUsers
      summary: Directory of instance users (invite and DM pickers).
      description: >
        Backs the two people-pickers the chat shell draws: "Invite people" on a
        channel and the "+" beside Direct messages. Returns the reduced
        UserSummary shape only — never email, role, or password state. Ordered
        by username ascending; the cursor encodes the last row's
        (username, id).
      security:
        - sessionCookie: []
      parameters:
        - name: q
          in: query
          description: Case-insensitive filter over username and display name.
          schema: { type: string, minLength: 1, maxLength: 64 }
        - name: limit
          in: query
          schema: { type: integer, minimum: 1, maximum: 100, default: 50 }
        - name: cursor
          in: query
          description: Opaque pagination cursor from a previous response.
          schema: { type: string }
      responses:
        "200":
          description: One page of the user directory.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserSummaryPage"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "429":
          $ref: "#/components/responses/RateLimited"
...
  /api/v1/admin/scim/tokens/{tokenId}:
    delete:
      tags: [admin]
      operationId: revokeScimToken
      summary: Revoke a provisioning token.
      description: >
        Takes effect immediately. The provider's next sync fails
        authentication, which is the intended way to cut off a system that
        should no longer be provisioning.
      security:
        - sessionCookie: []
      parameters:
        - name: tokenId
          in: path
          required: true
          schema: { type: string, format: uuid }
      responses:
        "204":
          description: The token is dead.
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
...
  /api/v1/admin/invites/{inviteId}:
    delete:
      operationId: revokeInvite
      summary: Revoke an open invite.
      description: >
        Idempotent: revoking one that is already gone answers 204, because
        the outcome the caller wanted is the outcome that holds.
      security:
        - sessionCookie: []
      parameters:
        - $ref: "#/components/parameters/InviteId"
      responses:
        "204":
          description: Revoked, or already gone.
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
...
components:
  securitySchemes:
    sessionCookie:
      type: apiKey
      in: cookie
      name: hamlaneh_session
      description: >
        HttpOnly+Secure+SameSite=Strict session cookie. Mutating requests must
        also send the X-Hamlaneh-CSRF header.
...
  responses:
    BadRequest:
      description: Request validation failed.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
...
  schemas:
...
    Error:
      type: object
      required: [error]
      properties:
        error:
          type: object
          required: [code, message]
          properties:
            code:
              type: string
              description: Stable machine-readable code (e.g. invalid_credentials).
            message:
              type: string
              description: Human-readable English message; clients localize by code.
...
    TotpLoginRequest:
      type: object
      required: [code]
      properties:
        code:
          type: string
          minLength: 6
          maxLength: 16
          description: >
            A six-digit authenticator code, or one recovery code
            (XXXX-XXXX; case-insensitive, hyphens and spaces ignored).
...